requirements.txt
flask==3.0.2
PyPDF2==3.0.1
pdf2image==1.16.3
python-docx==1.0.1
pillow==10.1.0
werkzeug==3.0.1
python-multipart==0.0.6
streamlit==1.28.2
pdf2docx==0.5.6 
PyMuPDF
docx2pdf==0.1.8